Split WAR.py into atomic files
