+++ src/neurodent/results/war_spikes.py
@@ -1,0 +1,189 @@
+"""Spike ingestion for :class:`WindowAnalysisResult` (issue #134)."""
+
+from __future__ import annotations
+
+import copy
+import logging
+from typing import Literal, TYPE_CHECKING
+
+import mne
+import numpy as np
+import pandas as pd
+
+from neurodent.core.utils import log_transform
+
+if TYPE_CHECKING:
+    from .frequency_domain_results import FrequencyDomainSpikeAnalysisResult
+    from .window_analysis_result import WindowAnalysisResult
+
+
+def _bin_spike_df(*args, **kwargs):
+    from .window_analysis_result import _bin_spike_df as _f
+    return _f(*args, **kwargs)
+
+class WARSpikeMixin:
+    """Mixin: see module docstring."""
+
+    def read_sars_spikes(
+        self,
+        sars: list["FrequencyDomainSpikeAnalysisResult"],
+        read_mode: Literal["sa", "mne"] = "sa",
+        inplace=True,
+    ):
+        """
+        Integrate spike analysis results into WAR by adding nspike/lognspike features.
+
+        This method extracts spike timing information from spike detection results and bins
+        them according to the WAR's time windows, adding spike count features to each row.
+
+        Args:
+            sars: List of FrequencyDomainSpikeAnalysisResult objects.
+                  One result per recording session (animalday).
+            read_mode: Mode for extracting spike data:
+                - "sa": Read from SortingAnalyzer objects (result_sas attribute)
+                - "mne": Read from MNE RawArray objects (result_mne attribute)
+            inplace: If True, modifies self.result and returns self.
+                    If False, returns a new WindowAnalysisResult.
+
+        Returns:
+            WindowAnalysisResult: WAR object with added spike features (nspike, lognspike).
+                - If inplace=True: returns self with modified result DataFrame
+                - If inplace=False: returns new WAR object with enhanced result DataFrame
+
+        Notes:
+            - The number of sars must match the number of unique animaldays in self.result
+            - Spikes are binned into time windows matching the existing WAR fragments
+            - nspike: array of spike counts per channel for each time window
+            - lognspike: log-transformed spike counts using log_transform()
+
+        Example:
+            >>> # After computing WAR and spike detection
+            >>> enhanced_war = war.read_sars_spikes(fdsar_list, read_mode="sa", inplace=False)
+            >>> enhanced_war.result['nspike']  # Spike counts per channel per window
+        """
+        match read_mode:
+            case "sa":
+                spikes_all = []
+                for sar in sars:  # for each continuous recording session
+                    spikes_channel = []
+                    for i, sa in enumerate(sar.result_sas):  # for each channel
+                        spike_times = []
+                        for unit in sa.sorting.get_unit_ids():  # Flatten units
+                            spike_times.extend(
+                                sa.sorting.get_unit_spike_train(unit_id=unit).tolist()
+                            )
+                        spike_times = (
+                            np.array(spike_times) / sa.sorting.get_sampling_frequency()
+                        )
+                        spikes_channel.append(spike_times)
+                    spikes_all.append(spikes_channel)
+                return self._read_from_spikes_all(spikes_all, inplace=inplace)
+            case "mne":
+                raws = [sar.result_mne for sar in sars]
+                return self.read_mnes_spikes(raws, inplace=inplace)
+            case _:
+                raise ValueError(f"Invalid read_mode: {read_mode}")
+
+    def read_mnes_spikes(self, raws: list[mne.io.RawArray], inplace=True):
+        """
+        Extract spike features from MNE RawArray objects with spike annotations.
+
+        This method extracts spike timing from MNE annotations (where spikes are marked
+        with channel-specific event labels) and bins them into WAR time windows.
+
+        Args:
+            raws: List of MNE RawArray objects with spike annotations. One per recording
+                  session (animalday). Each should have annotations with channel names
+                  as event labels (e.g., 'LMot', 'RMot', etc.).
+            inplace: If True, modifies self.result and returns self.
+                    If False, returns a new WindowAnalysisResult.
+
+        Returns:
+            WindowAnalysisResult: WAR object with added spike features (nspike, lognspike).
+
+        Notes:
+            - Expects MNE annotations with channel names as event descriptions
+            - Spike times are extracted from event onsets and binned to WAR windows
+            - Channels not found in annotations will have empty spike arrays
+            - Delegates to _read_from_spikes_all() for the actual binning logic
+
+        Example:
+            >>> # From MNE spike annotations
+            >>> enhanced_war = war.read_mnes_spikes([mne_raw1, mne_raw2], inplace=False)
+        """
+        spikes_all = []
+        for raw in raws:
+            # each mne is a contiguous recording session
+            events, event_id = mne.events_from_annotations(raw)
+            event_id = {k.item(): v for k, v in event_id.items()}
+
+            spikes_channel = []
+            for channel in raw.ch_names:
+                if channel not in event_id.keys():
+                    logging.warning(f"Channel {channel} not found in event_id")
+                    spikes_channel.append([])
+                    continue
+                event_id_channel = event_id[channel]
+                spike_times = events[events[:, 2] == event_id_channel, 0]
+                spike_times = spike_times / raw.info["sfreq"]
+                spikes_channel.append(spike_times)
+            spikes_all.append(spikes_channel)
+        return self._read_from_spikes_all(spikes_all, inplace=inplace)
+
+    def _read_from_spikes_all(self, spikes_all: list[list[list[float]]], inplace=True):
+        """
+        Internal method to bin spike times into WAR time windows and add as features.
+
+        This is the common endpoint for both read_sars_spikes() and read_mnes_spikes().
+        It bins spike times according to the WAR's time windows and adds nspike/lognspike
+        features to the result DataFrame.
+
+        Args:
+            spikes_all: Nested list structure of spike times in seconds:
+                - Outer list: recording sessions (one per animalday)
+                - Middle list: channels (one per EEG channel)
+                - Inner list/array: spike times in seconds for that channel
+                Example: [[[0.5, 1.2], [0.8]], [[1.1, 2.3], []]]
+                         = 2 sessions, 2 channels each
+            inplace: If True, modifies self.result and returns self.
+                    If False, returns a new WindowAnalysisResult with enhanced data.
+
+        Returns:
+            WindowAnalysisResult: WAR object with spike features added to result DataFrame.
+
+        Notes:
+            - Groups self.result by 'animalday' and matches to spikes_all by index
+            - Uses _bin_spike_df() helper to count spikes within each time window
+            - Adds two new columns:
+                - 'nspike': array of spike counts per channel for each window
+                - 'lognspike': log-transformed spike counts via log_transform()
+            - Warns if spike count size doesn't match result DataFrame size
+        """
+        # Each groupby animalday is a recording session
+        grouped = self.result.groupby("animalday")
+        animaldays = grouped.groups.keys()
+        logging.debug(f"Animal days: {animaldays}")
+        spike_counts = dict(zip(animaldays, spikes_all))
+        spike_counts = grouped.apply(
+            lambda x: _bin_spike_df(x, spikes_channel=spike_counts[x.name])
+        )
+        spike_counts: pd.Series = spike_counts.explode()
+
+        if spike_counts.size != self.result.shape[0]:
+            logging.warning(
+                f"Spike counts size {spike_counts.size} does not match result size {self.result.shape[0]}"
+            )
+
+        result = self.result.copy()
+        result["nspike"] = spike_counts.tolist()
+        result["lognspike"] = list(
+            log_transform(np.stack(result["nspike"].tolist(), axis=0))
+        )
+        if inplace:
+            self.result = result
+            return self
+        else:
+            # Create a new WindowAnalysisResult
+            new_war = copy.deepcopy(self)
+            new_war.result = result
+            return new_war
